Migrate conversion mirror tests to shared contract tests (Phase 1)
Moves the remaining 7 jdbc/r2dbc conversion mirror test pairs (CSV,
StringWrapper, NullReturningWritingConverter, CodeEnum,
ConverterRegistration, ValueClass, Collection) onto the shared contract
scaffold introduced in #445, following the same pattern as
EnumConversionContract.

ConverterContractBase gains an overridable `converters` hook so
contracts with a fixed converter set can register them without
duplicating the lazy client. CodeEnumConversionContract stays
independent of that base since it owns its own `code_enum` table.
CollectionConversionContract carries the 8 shared cases; the one
case where JDBC and R2DBC diverge (null element in an IN list) stays
on each concrete subclass, per the module-specific-case convention.

kuery-client-spring-data-testing now depends on spring-data-commons for
the converter annotations/interfaces the contracts register.
