+++ AGENTS.md
@@ -96,7 +96,7 @@ All modules apply `conventions.preset.base` (= `conventions.kotlin` + `conventio
 
 Behavior specified identically for both spring-data modules is written **once** as an abstract contract class in `kuery-client-spring-data-testing` (in its `main` source set, package `dev.hsbrysk.kuery.spring.testing.contract.<subpackage>`), against the suspending `KueryClient` API. Each module runs it via a concrete subclass in its own `src/test` with the same class name as before (e.g., `EnumConversionTest : EnumConversionContract`) that only supplies a `ContractDatabase` (jdbc adapts its blocking client via `BlockingKueryClientAdapter`). Module-specific extra cases go directly on the concrete subclass.
 
-Dependency rule: `kuery-client-spring-data-testing` depends only on `kuery-client-core` and test libraries — never on the jdbc/r2dbc implementations, so one module's test classpath is not polluted with the sibling implementation. `ContractDatabase` implementations (e.g., `JdbcH2ContractDatabase`) therefore live in each module's own `src/test`. Keep `ContractDatabase` itself minimal (client creation + raw `execute`); scenario-specific setup such as table DDL belongs to the contract classes.
+Dependency rule: `kuery-client-spring-data-testing` depends only on `kuery-client-core`, `spring-data-commons` (converter interfaces/annotations used by the conversion contracts), and test libraries — never on the jdbc/r2dbc implementations, so one module's test classpath is not polluted with the sibling implementation. `ContractDatabase` implementations (e.g., `JdbcH2ContractDatabase`) therefore live in each module's own `src/test`. Keep `ContractDatabase` itself minimal (client creation + raw `execute`); scenario-specific setup such as table DDL belongs to the contract classes.
 
 Contract scope for streaming: contracts may use `flow`/`flowMap` (jdbc runs them through the sequence-to-flow adapter) but must only assert on the **results** — values, mapping, order. Streaming **semantics** — lazy execution, cancellation, resource release, coroutine/Reactor context propagation, actual suspension — must not be asserted through the adapter; test them per module against the real API. Likewise, genuinely one-sided behavior (Reactor context on r2dbc, `Sequence` / `maxRows` / `queryTimeoutSeconds` on jdbc, transactions) stays as plain tests in the owning module.
