+++ kuery-client-spring-data-r2dbc/src/test/kotlin/dev/hsbrysk/kuery/spring/r2dbc/conversion/CSVConversionTest.kt
@@ -1,47 +1,12 @@
 package dev.hsbrysk.kuery.spring.r2dbc.conversion
 
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import dev.hsbrysk.kuery.core.single
-import dev.hsbrysk.kuery.spring.r2dbc.H2TestDatabase
-import kotlinx.coroutines.test.runTest
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
+import dev.hsbrysk.kuery.spring.r2dbc.R2dbcH2ContractDatabase
+import dev.hsbrysk.kuery.spring.testing.contract.conversion.CSVConversionContract
 
-class CSVConversionTest {
-    private val kueryClient = h2.kueryClient()
-
-    data class Record(val text: List<String>)
-
-    @BeforeEach
-    fun beforeEach() = runTest {
-        h2.setUpForConverterTest()
-    }
-
-    @AfterEach
-    fun afterEach() = runTest {
-        h2.tearDownForConverterTest()
-    }
-
-    @Test
-    fun `comma-separated text is read into a List property with trimmed elements`() = runTest {
-        // given
-        val text = "a, b,c"
-        kueryClient.sql {
-            +"INSERT INTO converter (text) VALUES ($text)"
-        }.rowsUpdated()
-
-        // when
-        val record: Record = kueryClient.sql {
-            +"SELECT * FROM converter"
-        }.single()
-
-        // then
-        assertThat(record.text).isEqualTo(listOf("a", "b", "c"))
-    }
+class CSVConversionTest : CSVConversionContract() {
+    override val database get() = h2
 
     companion object {
-        private val h2 = H2TestDatabase()
+        private val h2 = R2dbcH2ContractDatabase()
     }
 }
