+++ kuery-client-spring-data-r2dbc/src/test/kotlin/dev/hsbrysk/kuery/spring/r2dbc/conversion/CodeEnumConversionTest.kt
@@ -1,139 +1,12 @@
 package dev.hsbrysk.kuery.spring.r2dbc.conversion
 
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import dev.hsbrysk.kuery.core.single
-import dev.hsbrysk.kuery.spring.r2dbc.H2TestDatabase
-import kotlinx.coroutines.test.runTest
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.springframework.core.convert.converter.Converter
-import org.springframework.core.convert.converter.ConverterFactory
-import org.springframework.data.convert.ReadingConverter
-import org.springframework.data.convert.WritingConverter
-import org.springframework.r2dbc.core.awaitRowsUpdated
+import dev.hsbrysk.kuery.spring.r2dbc.R2dbcH2ContractDatabase
+import dev.hsbrysk.kuery.spring.testing.contract.conversion.CodeEnumConversionContract
 
-class CodeEnumConversionTest {
-    private val kueryClient = h2.kueryClient(
-        listOf(
-            IntCodeEnumWritingConverter(),
-            IntCodeEnumReadingConverter(),
-            StringCodeEnumWritingConverter(),
-            StringCodeEnumReadingConverter(),
-        ),
-    )
-
-    @BeforeEach
-    fun beforeEach() = runTest {
-        h2.databaseClient.sql(
-            """
-            CREATE TABLE code_enum
-            (
-                id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-                int_enum INT DEFAULT NULL,
-                string_enum VARCHAR(255) DEFAULT NULL
-            )
-            """.trimIndent(),
-        ).fetch().awaitRowsUpdated()
-    }
-
-    @AfterEach
-    fun afterEach() = runTest {
-        h2.databaseClient.sql("DROP TABLE code_enum").fetch().awaitRowsUpdated()
-    }
-
-    interface CodeEnum<T> {
-        val code: T
-
-        companion object {
-            fun <T, E : CodeEnum<T>> getByCode(
-                code: T,
-                clazz: Class<E>,
-            ): E? = clazz.enumConstants.firstOrNull { code == it.code }
-        }
-    }
-
-    interface IntCodeEnum : CodeEnum<Int>
-
-    interface StringCodeEnum : CodeEnum<String>
-
-    enum class SampleIntCodeEnum(override val code: Int) : IntCodeEnum {
-        HOGE(10),
-    }
-
-    enum class SampleStringCodeEnum(override val code: String) : StringCodeEnum {
-        BAR("hoge"),
-    }
-
-    data class Record(
-        val intEnum: SampleIntCodeEnum,
-        val stringEnum: SampleStringCodeEnum,
-    )
-
-    @WritingConverter
-    class IntCodeEnumWritingConverter : Converter<IntCodeEnum, Int> {
-        override fun convert(source: IntCodeEnum): Int = source.code
-    }
-
-    @ReadingConverter
-    class IntCodeEnumReadingConverter : ConverterFactory<Int, IntCodeEnum> {
-        override fun <E : IntCodeEnum> getConverter(targetType: Class<E>): Converter<Int, E?> = Converter {
-            CodeEnum.getByCode(it, targetType)
-        }
-    }
-
-    @WritingConverter
-    class StringCodeEnumWritingConverter : Converter<StringCodeEnum, String> {
-        override fun convert(source: StringCodeEnum): String = source.code
-    }
-
-    @ReadingConverter
-    class StringCodeEnumReadingConverter : ConverterFactory<String, StringCodeEnum> {
-        override fun <E : StringCodeEnum> getConverter(targetType: Class<E>): Converter<String, E?> = Converter {
-            CodeEnum.getByCode(it, targetType)
-        }
-    }
-
-    @Test
-    fun `code enums are written as their codes and mapped back to enum properties`() = runTest {
-        // given
-        kueryClient.sql {
-            +"INSERT INTO code_enum (int_enum, string_enum)"
-            +"VALUES (${SampleIntCodeEnum.HOGE}, ${SampleStringCodeEnum.BAR})"
-        }.rowsUpdated()
-
-        // when
-        val record: Record = kueryClient.sql {
-            +"SELECT * FROM code_enum"
-        }.single()
-
-        // then
-        assertThat(record.intEnum).isEqualTo(SampleIntCodeEnum.HOGE)
-        assertThat(record.stringEnum).isEqualTo(SampleStringCodeEnum.BAR)
-    }
-
-    @Test
-    fun `code enum is converted when fetched as a scalar`() = runTest {
-        // given
-        kueryClient.sql {
-            +"INSERT INTO code_enum (int_enum, string_enum)"
-            +"VALUES (${SampleIntCodeEnum.HOGE}, ${SampleStringCodeEnum.BAR})"
-        }.rowsUpdated()
-
-        // when & then
-        val intEnum: SampleIntCodeEnum = kueryClient.sql {
-            +"SELECT int_enum FROM code_enum"
-        }.single()
-        assertThat(intEnum).isEqualTo(SampleIntCodeEnum.HOGE)
-
-        val stringEnum: SampleStringCodeEnum = kueryClient.sql {
-            +"SELECT string_enum FROM code_enum"
-        }.single()
-        assertThat(stringEnum).isEqualTo(SampleStringCodeEnum.BAR)
-    }
+class CodeEnumConversionTest : CodeEnumConversionContract() {
+    override val database get() = h2
 
     companion object {
-        private val h2 = H2TestDatabase()
+        private val h2 = R2dbcH2ContractDatabase()
     }
 }
