+++ kuery-client-spring-data-r2dbc/src/test/kotlin/dev/hsbrysk/kuery/spring/r2dbc/conversion/StringWrapperConversionTest.kt
@@ -1,66 +1,12 @@
 package dev.hsbrysk.kuery.spring.r2dbc.conversion
 
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import dev.hsbrysk.kuery.core.single
-import dev.hsbrysk.kuery.spring.r2dbc.H2TestDatabase
-import kotlinx.coroutines.test.runTest
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.springframework.core.convert.converter.Converter
-import org.springframework.data.convert.ReadingConverter
-import org.springframework.data.convert.WritingConverter
+import dev.hsbrysk.kuery.spring.r2dbc.R2dbcH2ContractDatabase
+import dev.hsbrysk.kuery.spring.testing.contract.conversion.StringWrapperConversionContract
 
-class StringWrapperConversionTest {
-    private val kueryClient = h2.kueryClient(
-        listOf(
-            StringWrapperToStringConverter(),
-            StringToStringWrapperConverter(),
-        ),
-    )
-
-    data class StringWrapper(val value: String)
-
-    data class Record(val text: StringWrapper)
-
-    @WritingConverter
-    class StringWrapperToStringConverter : Converter<StringWrapper, String> {
-        override fun convert(source: StringWrapper): String = source.value
-    }
-
-    @ReadingConverter
-    class StringToStringWrapperConverter : Converter<String, StringWrapper> {
-        override fun convert(source: String): StringWrapper = StringWrapper(source)
-    }
-
-    @BeforeEach
-    fun beforeEach() = runTest {
-        h2.setUpForConverterTest()
-    }
-
-    @AfterEach
-    fun afterEach() = runTest {
-        h2.tearDownForConverterTest()
-    }
-
-    @Test
-    fun `wrapper type round-trips through registered writing and reading converters`() = runTest {
-        // given
-        kueryClient.sql {
-            +"INSERT INTO converter (text) VALUES (${StringWrapper("hoge")})"
-        }.rowsUpdated()
-
-        // when
-        val record: Record = kueryClient.sql {
-            +"SELECT * FROM converter"
-        }.single()
-
-        // then
-        assertThat(record.text).isEqualTo(StringWrapper("hoge"))
-    }
+class StringWrapperConversionTest : StringWrapperConversionContract() {
+    override val database get() = h2
 
     companion object {
-        private val h2 = H2TestDatabase()
+        private val h2 = R2dbcH2ContractDatabase()
     }
 }
