+++ kuery-client-spring-data-testing/src/main/kotlin/dev/hsbrysk/kuery/spring/testing/contract/conversion/CodeEnumConversionContract.kt
@@ -1,0 +1,147 @@
+package dev.hsbrysk.kuery.spring.testing.contract.conversion
+
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import dev.hsbrysk.kuery.core.KueryClient
+import dev.hsbrysk.kuery.core.single
+import dev.hsbrysk.kuery.spring.testing.ContractDatabase
+import kotlinx.coroutines.test.runTest
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.springframework.core.convert.converter.Converter
+import org.springframework.core.convert.converter.ConverterFactory
+import org.springframework.data.convert.ReadingConverter
+import org.springframework.data.convert.WritingConverter
+
+/**
+ * Contract shared by the jdbc and r2dbc modules. Each module runs it via a concrete subclass
+ * that supplies its `ContractDatabase`.
+ *
+ * Uses its own `code_enum` table rather than [ConverterContractBase]'s shared `converter` table,
+ * so it does not extend that base.
+ */
+abstract class CodeEnumConversionContract {
+    protected abstract val database: ContractDatabase
+
+    // lazy: `database` is not resolvable yet while this base class initializes.
+    protected val kueryClient: KueryClient by lazy {
+        database.kueryClient(
+            listOf(
+                IntCodeEnumWritingConverter(),
+                IntCodeEnumReadingConverter(),
+                StringCodeEnumWritingConverter(),
+                StringCodeEnumReadingConverter(),
+            ),
+        )
+    }
+
+    @BeforeEach
+    fun setUpCodeEnumTable() {
+        database.execute(
+            """
+            CREATE TABLE code_enum
+            (
+                id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+                int_enum INT DEFAULT NULL,
+                string_enum VARCHAR(255) DEFAULT NULL
+            )
+            """.trimIndent(),
+        )
+    }
+
+    @AfterEach
+    fun dropCodeEnumTable() {
+        database.execute("DROP TABLE code_enum")
+    }
+
+    interface CodeEnum<T> {
+        val code: T
+
+        companion object {
+            fun <T, E : CodeEnum<T>> getByCode(
+                code: T,
+                clazz: Class<E>,
+            ): E? = clazz.enumConstants.firstOrNull { code == it.code }
+        }
+    }
+
+    interface IntCodeEnum : CodeEnum<Int>
+
+    interface StringCodeEnum : CodeEnum<String>
+
+    enum class SampleIntCodeEnum(override val code: Int) : IntCodeEnum {
+        HOGE(10),
+    }
+
+    enum class SampleStringCodeEnum(override val code: String) : StringCodeEnum {
+        BAR("hoge"),
+    }
+
+    data class Record(
+        val intEnum: SampleIntCodeEnum,
+        val stringEnum: SampleStringCodeEnum,
+    )
+
+    @WritingConverter
+    class IntCodeEnumWritingConverter : Converter<IntCodeEnum, Int> {
+        override fun convert(source: IntCodeEnum): Int = source.code
+    }
+
+    @ReadingConverter
+    class IntCodeEnumReadingConverter : ConverterFactory<Int, IntCodeEnum> {
+        override fun <E : IntCodeEnum> getConverter(targetType: Class<E>): Converter<Int, E?> = Converter {
+            CodeEnum.getByCode(it, targetType)
+        }
+    }
+
+    @WritingConverter
+    class StringCodeEnumWritingConverter : Converter<StringCodeEnum, String> {
+        override fun convert(source: StringCodeEnum): String = source.code
+    }
+
+    @ReadingConverter
+    class StringCodeEnumReadingConverter : ConverterFactory<String, StringCodeEnum> {
+        override fun <E : StringCodeEnum> getConverter(targetType: Class<E>): Converter<String, E?> = Converter {
+            CodeEnum.getByCode(it, targetType)
+        }
+    }
+
+    @Test
+    fun `code enums are written as their codes and mapped back to enum properties`() = runTest {
+        // given
+        kueryClient.sql {
+            +"INSERT INTO code_enum (int_enum, string_enum)"
+            +"VALUES (${SampleIntCodeEnum.HOGE}, ${SampleStringCodeEnum.BAR})"
+        }.rowsUpdated()
+
+        // when
+        val record: Record = kueryClient.sql {
+            +"SELECT * FROM code_enum"
+        }.single()
+
+        // then
+        assertThat(record.intEnum).isEqualTo(SampleIntCodeEnum.HOGE)
+        assertThat(record.stringEnum).isEqualTo(SampleStringCodeEnum.BAR)
+    }
+
+    @Test
+    fun `code enum is converted when fetched as a scalar`() = runTest {
+        // given
+        kueryClient.sql {
+            +"INSERT INTO code_enum (int_enum, string_enum)"
+            +"VALUES (${SampleIntCodeEnum.HOGE}, ${SampleStringCodeEnum.BAR})"
+        }.rowsUpdated()
+
+        // when & then
+        val intEnum: SampleIntCodeEnum = kueryClient.sql {
+            +"SELECT int_enum FROM code_enum"
+        }.single()
+        assertThat(intEnum).isEqualTo(SampleIntCodeEnum.HOGE)
+
+        val stringEnum: SampleStringCodeEnum = kueryClient.sql {
+            +"SELECT string_enum FROM code_enum"
+        }.single()
+        assertThat(stringEnum).isEqualTo(SampleStringCodeEnum.BAR)
+    }
+}
